Move dl Unmarshal to the es.HitT type
